requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
fpdf2
xlsxwriter
